Fixed several issues
FIXED: Major issue in which the game would freeze if units were moved in a
certain way, or in certain conditions.

FIXED: Major issue with loading in which units were loaded improperly,
this also means that a change in the format was done again.

FIXED: Minor issue in which unit mage2Player (unitID 1 (previously unitID
2)) was set for both the correct unit and warrior1Player (unitID 2
(previously unitID 3))